feat: move admin page to mon compte page (#1631)
* feat: move admin page to mon compte page

* content typo

* feat: create a dedicated api route for admin

* refactor: use a dedicated /admin page

* feat: metadata and seo

* Update app/(header-compte)/admin/page.tsx


* Update app/(header-compte)/admin/page.tsx


* Update page.tsx


---------
